# Index spree_orders on various columns