--- a/db/migrate/20190916105416_add_long_compound_index_on_spree_orders.rb
+++ b/db/migrate/20190916105416_add_long_compound_index_on_spree_orders.rb
@@ -1,0 +1,9 @@
+class AddLongCompoundIndexOnSpreeOrders < ActiveRecord::Migration
+  def change
+    add_index(
+      :spree_orders,
+      [:completed_at, :user_id, :created_by_id, :created_at],
+      name: 'spree_orders_completed_at_user_id_created_by_id_created_at_idx'
+    )
+  end
+end