--- a/db/migrate/20190916110029_drop_completed_at_index_on_spree_orders.rb
+++ b/db/migrate/20190916110029_drop_completed_at_index_on_spree_orders.rb
@@ -1,0 +1,5 @@
+class DropCompletedAtIndexOnSpreeOrders < ActiveRecord::Migration
+  def change
+    remove_index :spree_orders, :completed_at
+  end
+end