@sauloperez 2019-09-16
> I say this because this is a query specific index so I think it deserves a comment where the query is

It's actually used in many places. Take a look at https://app.datadoghq.com/apm/resource/rails-postgres/postgres.query/9b4df5714e78b76f?end=1568637056648&env=production&paused=false&start=1566045056648&index=apm-search

> If we change the query in the future, we should at least verify this index as well.

We do have the monitoring in place to watch for that so as long as we pay attention to Datadog we should be covered.