@luisramos0 2019-09-16
I am ok with that but the calls listed are the controllers, there must be a common ground to these calls, some operation on a specific model or something.