@sauloperez 2019-09-17
Alright, so this query is triggered by `#last_incomplete_spree_order` which is defined in spree in [core/config/initializers/user_class_extensions.rb](https://github.com/openfoodfoundation/spree/blob/master/core/config/initializers/user_class_extensions.rb) 

```ruby
def last_incomplete_spree_order
  spree_orders.incomplete.where(:created_by_id => self.id).order('created_at DESC').first
end
```

which in turn is used by OFN in https://github.com/openfoodfoundation/openfoodnetwork/blob/07ae51560e9063318fe388b428be6e3fd7d8869a/lib/spree/core/controller_helpers/order_decorator.rb#L20-L30

pretty much used everywhere.