@luisramos0 2019-09-17
ah, yeah, thanks for clarifying that.
and that is a before_filter in the store_controller, basically, this is called in every frontoffice action 👍