@sauloperez 2019-09-17
Please, stop touching the db/schema :trollface: