@sauloperez 2019-09-18
I see now that for this case even a smoke test doesn't make sense. A DB can't affect any functionality.

This can only be deployed at most to check if the migrations work.