@sauloperez 2019-09-19
Successfully deployed and migrated in fr-staging.